+++ src/path.h
@@ -23,4 +23,11 @@ char_u *concat_str(char_u *str1, char_u *str2);
 void add_pathsep(char_u *p);
 char_u *FullName_save(char_u *fname, int force);
 void simplify_filename(char_u *filename);
+char_u *find_file_name_in_path(char_u *ptr, int len, int options,
+                               long count,
+                               char_u *rel_fname);
+int path_is_url(char_u *p);
+int path_with_url(char_u *fname);
+int vim_isAbsName(char_u *name);
+int vim_FullName(char_u *fname, char_u *buf, int len, int force);
 #endif
